LayoutTests/inspector/debugger/live-edit-expected.txt

CONSOLE MESSAGE: line 3: Uncaught SyntaxError: Unexpected token ,
Tests live edit feature.

Debugger was enabled.

Running: testLiveEdit
==Source frame contents start==
function f()
{
    return "live-edited string";
}

==Source frame contents end==

Running: testLiveEditSyntaxError
==Source frame contents start==
function syntaxError()
{
    return "I'm good";
}

==Source frame contents end==

Running: testLiveEditWhenPaused
Script execution paused.
Script execution paused.
Script execution resumed.

Running: testNoCrashWhenOnlyOneFunctionOnStack
Script execution paused.
Call stack:
    0) f1 (edit-me-when-paused.js:13)
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testBreakpointsUpdated
Debugger was disabled.
